fixed Col number in formulas
