+++ app/components/Sheet/ColumnEditMenu/FormulaMenuItem.jsx
@@ -8,11 +8,15 @@ import ContentEditable from 'react-contenteditable';
 const cx = classNames.bind(styles);
 
 const FormulaMenuItem = (props) => {
+	// componenet will mount fetachformals 
+	// if the form is already loaded then don't fetch?
 	return (
 	    <div className='col-md-12'>
-			<p className='col-md-12'>Allows you to create custom formulas for manipulating your data.</p>
-			<ContentEditable className='clearfix col-md-12' onChange={props.handleFormulaNameChange} html={props.formulaName||'Please Name'} /> <br></br>
-
+			<h5 className='col-md-12'>Allows you to create custom formulas for manipulating your data.</h5>
+			<div className='row clearfix'>
+				<label className='col-md-4' for='handleFormulaNameChange'>Name:</label>
+				<ContentEditable className='col-md-8' id='handleFormulaNameChange' style={{backgroundColor: 'white'}} onChange={props.handleFormulaNameChange} html={props.formulaName} /> 
+			</div>
 			<textarea onChange={props.handleFormulaCustom} className='col-md-12' value={props.formula} />
 			<button className="btn col-md-8 col-md-offset-4" type="button" onClick={props.formulaUpload}>Upload Formula</button>
 		</div>
